+++ app/src/main/java/com/undef/localhandsbrambillafunes/data/local/db/Converters.kt
@@ -1,4 +1,4 @@
-package com.undef.localhandsbrambillafunes.data.model.db
+package com.undef.localhandsbrambillafunes.data.local.db
 
 import androidx.room.TypeConverter
