+++ app/src/main/java/com/undef/localhandsbrambillafunes/data/local/entities/User.kt
@@ -1,4 +1,4 @@
-package com.undef.localhandsbrambillafunes.data.model.entities
+package com.undef.localhandsbrambillafunes.data.local.entities
 
 import androidx.room.Entity
 import androidx.room.PrimaryKey
