+++ app/src/main/java/com/undef/localhandsbrambillafunes/data/local/viewmodel/SessionViewModel.kt
@@ -1,10 +1,10 @@
-package com.undef.localhandsbrambillafunes.data.model.viewmodel
+package com.undef.localhandsbrambillafunes.data.local.viewmodel
 
 import android.app.Application
 import androidx.lifecycle.AndroidViewModel
 import androidx.lifecycle.viewModelScope
-import com.undef.localhandsbrambillafunes.data.model.entities.User
-import com.undef.localhandsbrambillafunes.data.repository.UserRepository
+import com.undef.localhandsbrambillafunes.data.local.entities.User
+import com.undef.localhandsbrambillafunes.data.local.repository.UserRepository
 import kotlinx.coroutines.flow.MutableStateFlow
 import kotlinx.coroutines.flow.StateFlow
 import kotlinx.coroutines.launch
