+++ app/src/main/java/com/undef/localhandsbrambillafunes/ui/screens/entrepreneur/ProductOwnerDetailScreen.kt
@@ -49,7 +49,7 @@ import androidx.compose.ui.text.style.TextAlign
 import androidx.compose.ui.unit.dp
 import androidx.navigation.NavController
 import coil.compose.AsyncImage
-import com.undef.localhandsbrambillafunes.data.model.entities.Product
+import com.undef.localhandsbrambillafunes.data.local.entities.Product
 import com.undef.localhandsbrambillafunes.ui.navigation.AppScreens
 
 /**
