+++ app/src/main/java/com/undef/localhandsbrambillafunes/ui/screens/entrepreneur/SellScreen.kt
@@ -40,9 +40,8 @@ import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.unit.dp
-import androidx.lifecycle.viewmodel.compose.viewModel
 import androidx.navigation.NavController
-import com.undef.localhandsbrambillafunes.data.model.viewmodel.ProductViewModel
+import com.undef.localhandsbrambillafunes.data.local.viewmodel.ProductViewModel
 import com.undef.localhandsbrambillafunes.ui.navigation.AppScreens
 import androidx.compose.runtime.getValue
 import androidx.compose.foundation.lazy.items
@@ -51,7 +50,7 @@ import androidx.compose.ui.draw.clip
 import androidx.compose.ui.layout.ContentScale
 import androidx.compose.ui.text.font.FontWeight
 import coil.compose.rememberAsyncImagePainter
-import com.undef.localhandsbrambillafunes.data.model.viewmodel.SessionViewModel
+import com.undef.localhandsbrambillafunes.data.local.viewmodel.SessionViewModel
 
 
 /**
