+++ app/src/main/java/com/undef/localhandsbrambillafunes/ui/screens/home/components/SearchBarScreen.kt
@@ -40,9 +40,9 @@ import androidx.compose.ui.Modifier
 import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.unit.dp
 import androidx.navigation.NavController
-import com.undef.localhandsbrambillafunes.data.model.entities.Product
-import com.undef.localhandsbrambillafunes.data.model.ProductListItem
-import com.undef.localhandsbrambillafunes.data.model.ProductProvider
+import com.undef.localhandsbrambillafunes.data.local.entities.Product
+import com.undef.localhandsbrambillafunes.data.local.model.ProductListItem
+import com.undef.localhandsbrambillafunes.data.local.model.ProductProvider
 import com.undef.localhandsbrambillafunes.ui.navigation.AppScreens
 
 /**
